move init of plugin lib to init_plugins.py
