+++ mmcv/tensorrt/__init__.py
@@ -1,9 +1,12 @@
 # flake8: noqa
-from .tensorrt_utils import (TRTWraper, get_tensorrt_op_path,
-                             is_tensorrt_plugin_loaded, load_tensorrt_plugin,
-                             load_trt_engine, onnx2trt, save_trt_engine)
+from .init_plugins import is_tensorrt_plugin_loaded, load_tensorrt_plugin
+from .tensorrt_utils import (TRTWraper, load_trt_engine, onnx2trt,
+                             save_trt_engine)
+
+# load tensorrt plugin lib
+load_tensorrt_plugin()
 
 __all__ = [
-    'onnx2trt', 'save_trt_engine', 'load_trt_engine', 'get_tensorrt_op_path',
-    'is_tensorrt_plugin_loaded', 'load_tensorrt_plugin', 'TRTWraper'
+    'onnx2trt', 'save_trt_engine', 'load_trt_engine', 'TRTWraper',
+    'is_tensorrt_plugin_loaded'
 ]
