+++ mmcv/tensorrt/tensorrt_utils.py
@@ -1,43 +1,5 @@
-import ctypes
-import glob
-import os
-
 import tensorrt as trt
 import torch
-
-
-def get_tensorrt_op_path():
-    """Get TensorRT plugins library path."""
-    wildcard = os.path.join(
-        os.path.abspath(os.path.dirname(os.path.dirname(__file__))),
-        '_ext_trt.*.so')
-
-    paths = glob.glob(wildcard)
-    if len(paths) > 0:
-        return paths[0]
-    else:
-        return None
-
-
-plugin_is_loaded = False
-
-
-def is_tensorrt_plugin_loaded():
-    """Check if TensorRT plugins library is loaded or not.
-
-    Returns:
-        bool: plugin_is_loaded flag
-    """
-    global plugin_is_loaded
-    return plugin_is_loaded
-
-
-def load_tensorrt_plugin():
-    """load TensorRT plugins library."""
-    global plugin_is_loaded
-    if not plugin_is_loaded:
-        ctypes.CDLL(get_tensorrt_op_path())
-        plugin_is_loaded = True
 
 
 def onnx2trt(onnx_model,
@@ -72,9 +34,6 @@ def onnx2trt(onnx_model,
         >>>             device_id=0)
         >>>             })
     """
-    # import plugin
-    load_tensorrt_plugin()
-
     device = torch.device('cuda:{}'.format(device_id))
     # create builder and network
     logger = trt.Logger(log_level)
@@ -185,7 +144,6 @@ class TRTWraper(torch.nn.Module):
 
     def __init__(self, engine=None, input_names=None, output_names=None):
         super(TRTWraper, self).__init__()
-        load_tensorrt_plugin()
         self._register_state_dict_hook(TRTWraper._on_state_dict)
         self.engine = engine
         if isinstance(self.engine, str):
